requirements.txt
asgiref==3.8.1
attrs==25.4.0
aws-sns-message-validator==0.0.5
boto3==1.40.50
botocore==1.40.50
braintree==4.33.1
brotli==1.2.0
CacheControl==0.14.2
cachetools==5.5.2
cattrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
cryptography==44.0.2
Django==4.2.18
django-cors-headers==4.4.0
djangorestframework==3.15.2
exceptiongroup==1.3.0
firebase-admin==6.7.0
fonttools==4.60.1
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
googletrans==4.0.0rc1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jmespath==1.0.1
msgpack==1.1.0
oauthlib==3.3.1
packaging==25.0
platformdirs==4.4.0
polib==1.2.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.32.3
requests-cache==1.2.1
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
url-normalize==2.2.1
urllib3==2.3.0
whitenoise==6.11.0
